Binary saving moved to the .c file.
Both binary-binary and binary-hdf5 tests pass.
