+++ src/core_save.h
@@ -12,9 +12,7 @@ extern "C" {
     extern void save_galaxies(const int filenr, const int tree, const int numgals, struct halo_data *halos,
                               struct halo_aux_data *haloaux, struct GALAXY *halogal, struct save_info *save_info,
                               const struct params *run_params);
-    extern void prepare_galaxy_for_output(int filenr, int tree, struct GALAXY *g, struct GALAXY_OUTPUT *o, struct halo_data *halos,
-                                          struct halo_aux_data *haloaux, struct GALAXY *halogal, const struct params *run_params);
-    extern int finalize_galaxy_file(const int ntrees, struct save_info *save_info, const struct params *run_params);
+    extern int finalize_galaxy_files(const int ntrees, struct save_info *save_info, const struct params *run_params);
     
 #ifdef __cplusplus
 }
